moved preset-messages into nltools2
