+++ projects/epc/playground/src/presets/EditBufferMockFactory.h
@@ -1,14 +1,15 @@
 #pragma once
 #ifdef _DEVELOPMENT_PC
 #include <nltools/messaging/Message.h>
+#include "PresetMessages.h"
 
 class EditBufferMockFactory
 {
  public:
-  static nltools::msg::SinglePresetMessage initialMessage();
-  static nltools::msg::SplitPresetMessage splitMessage();
-  static nltools::msg::SinglePresetMessage splitToSingleMessage();
-  static nltools::msg::LayerPresetMessage layerMessage();
-  static nltools::msg::SinglePresetMessage layerToSingleMessage();
+  static nltools2::msg::SinglePresetMessage initialMessage();
+  static nltools2::msg::SplitPresetMessage splitMessage();
+  static nltools2::msg::SinglePresetMessage splitToSingleMessage();
+  static nltools2::msg::LayerPresetMessage layerMessage();
+  static nltools2::msg::SinglePresetMessage layerToSingleMessage();
 };
 #endif
